+++ CHANGES.md
@@ -13,6 +13,7 @@
 - Android startOnBoot option
 - Android startForeground option
 - Android http posting of locations (options url and httpHeaders)
+- iOS showLocationSettings
 
 #### Removed
 - WP8 platform